research_agent/requirements.txt
crewai==0.41.1
crewai-tools==0.4.0
groq==0.9.0
tavily-python==0.3.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
arxiv==2.1.0
